+++ source/extensions/omni.isaac.examples/omni/isaac/examples/robo_party/robo_party.py
@@ -9,9 +9,6 @@
 
 import carb
 import numpy as np
-from omni.isaac.core.prims.xform_prim import XFormPrim
-from omni.isaac.dofbot.controllers import PickPlaceController
-from omni.isaac.dofbot.tasks import PickPlace
 from omni.isaac.examples.base_sample import BaseSample
 from omni.isaac.franka.controllers.stacking_controller import StackingController as FrankaStackingController
 from omni.isaac.franka.tasks import Stacking as FrankaStacking
@@ -30,7 +27,6 @@ class RoboParty(BaseSample):
         self._tasks = []
         self._controllers = []
         self._articulation_controllers = []
-        self._pick_place_task_params = None
         self._robots = []
         return
 
@@ -39,8 +35,6 @@ class RoboParty(BaseSample):
         self._tasks.append(FrankaStacking(name="task_0", offset=np.array([0, -2, 0])))
         world.add_task(self._tasks[-1])
         self._tasks.append(UR10Stacking(name="task_1", offset=np.array([0.5, 0.5, 0])))
-        world.add_task(self._tasks[-1])
-        self._tasks.append(PickPlace(name="task_2", offset=np.array([0, -1, 0])))
         world.add_task(self._tasks[-1])
 
         assets_root_path = get_assets_root_path()
@@ -77,13 +71,11 @@ class RoboParty(BaseSample):
         self._tasks = [
             self._world.get_task(name="task_0"),
             self._world.get_task(name="task_1"),
-            self._world.get_task(name="task_2"),
         ]
-        for i in range(3):
+        for i in range(2):
             self._robots.append(self._world.scene.get_object(self._tasks[i].get_params()["robot_name"]["value"]))
         self._robots.append(self._world.scene.get_object("my_kaya"))
         self._robots.append(self._world.scene.get_object("my_jetbot"))
-        self._pick_place_task_params = self._tasks[2].get_params()
         self._controllers.append(
             FrankaStackingController(
                 name="stacking_controller",
@@ -100,11 +92,6 @@ class RoboParty(BaseSample):
                 robot_articulation=self._robots[1],
                 picking_order_cube_names=self._tasks[1].get_cube_names(),
                 robot_observation_name=self._robots[1].name,
-            )
-        )
-        self._controllers.append(
-            PickPlaceController(
-                name="pick_place_controller", gripper=self._robots[2].gripper, robot_articulation=self._robots[2]
             )
         )
 
@@ -131,7 +118,7 @@ class RoboParty(BaseSample):
             )
         )
         self._controllers.append(DifferentialController(name="simple_control", wheel_radius=0.03, wheel_base=0.1125))
-        for i in range(5):
+        for i in range(4):
             self._articulation_controllers.append(self._robots[i].get_articulation_controller())
         return
 
@@ -141,25 +128,15 @@ class RoboParty(BaseSample):
         self._articulation_controllers[0].apply_action(actions)
         actions = self._controllers[1].forward(observations=observations, end_effector_offset=np.array([0, 0, 0.02]))
         self._articulation_controllers[1].apply_action(actions)
-
-        actions = self._controllers[2].forward(
-            picking_position=observations[self._pick_place_task_params["cube_name"]["value"]]["position"],
-            placing_position=observations[self._pick_place_task_params["cube_name"]["value"]]["target_position"],
-            current_joint_positions=observations[self._pick_place_task_params["robot_name"]["value"]][
-                "joint_positions"
-            ],
-            end_effector_offset=np.array([0, -0.06, 0]),
-        )
-        self._articulation_controllers[2].apply_action(actions)
         if self._world.current_time_step_index >= 0 and self._world.current_time_step_index < 500:
-            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0.2, 0.0, 0.0]))
-            self._robots[4].apply_wheel_actions(self._controllers[4].forward(command=[0.1, 0]))
+            self._robots[2].apply_wheel_actions(self._controllers[2].forward(command=[0.2, 0.0, 0.0]))
+            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0.1, 0]))
         elif self._world.current_time_step_index >= 500 and self._world.current_time_step_index < 1000:
-            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0, 0.2, 0.0]))
-            self._robots[4].apply_wheel_actions(self._controllers[4].forward(command=[0.0, np.pi / 10]))
+            self._robots[2].apply_wheel_actions(self._controllers[2].forward(command=[0, 0.2, 0.0]))
+            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0.0, np.pi / 10]))
         elif self._world.current_time_step_index >= 1000 and self._world.current_time_step_index < 1500:
-            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0, 0.0, 0.06]))
-            self._robots[4].apply_wheel_actions(self._controllers[4].forward(command=[0.1, 0]))
+            self._robots[2].apply_wheel_actions(self._controllers[2].forward(command=[0, 0.0, 0.06]))
+            self._robots[3].apply_wheel_actions(self._controllers[3].forward(command=[0.1, 0]))
         return
 
     async def _on_start_party_event_async(self):
@@ -180,6 +157,5 @@ class RoboParty(BaseSample):
         self._tasks = []
         self._controllers = []
         self._articulation_controllers = []
-        self._pick_place_task_params = None
         self._robots = []
         return