+++ source/extensions/omni.isaac.tests/omni/isaac/tests/tests/test_amr_transporter.py
@@ -81,7 +81,7 @@ class TestIw_hub(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        # get the dofbot
+        # get the jetbot
         self.ar = Articulation("/iw_hub")
         self.ar._articulation_view.initialize()
         self.starting_pos, _ = self.ar.get_world_pose()