+++ source/extensions/omni.isaac.motion_generation/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [7.1.0] - 2024-07-10
+### Removed
+- Deprecated omni.isaac.dofbot and removed its usage.
+
 ## [7.0.2] - 2024-06-12
 ### Changed
 - No longer override Franka gains in RRT tests.  These tests now serve as a check to see if the Franka can closely follow a trajectory with its default gains.